Welcome and Tour fix
Fixed issue where tour runs before welcome screen dismissed / vitals screen is active
