+++ Project MyFi/js/dashboard.js
@@ -5,6 +5,7 @@ import { initHUD } from "./hud/hud.js";
 import { createSplash } from './splash.js';
 
 import { updateIncome, updateCoreExpenses } from "./data/cashflowData.js";
+import { showWelcomeThenMaybeSetup } from './welcome.js';
 
 import "./core/truelayer.js";
 import "./modal.js";
@@ -60,254 +61,28 @@ document.addEventListener("DOMContentLoaded", () => {
       return d;
     }
 
-    async function showFirstRunSetup(uid) {
-      const userRef = doc(db, 'players', uid);
-      const snap = await getDoc(userRef);
-      const alreadyOnboarded = snap.exists() && !!snap.data()?.onboardedAt;
-      if (alreadyOnboarded) return;
-
-      const shell = document.createElement('div');
-      shell.id = 'welcomeOverlay';
-      Object.assign(shell.style, {
-        position: 'fixed', inset: '0', display: 'grid', placeItems: 'center',
-        padding: '16px', background: 'rgba(10,10,14,.65)', backdropFilter: 'blur(6px)',
-        zIndex: '9999', width: '100%', maxWidth: '100%', overflowX: 'hidden', boxSizing: 'border-box'
-      });
-
-      const card = document.createElement('div');
-      Object.assign(card.style, {
-        width: 'min(560px, 94vw)', maxHeight: '90vh', overflow: 'auto',
-        background: 'rgba(22,22,28,.96)', border: '1px solid var(--line, #333)',
-        borderRadius: '16px', padding: '18px 16px 12px',
-        boxShadow: '0 18px 50px rgba(0,0,0,.35)', boxSizing: 'border-box'
-      });
-
-      const title = document.createElement('h1'); title.textContent = 'Quick Setup'; title.style.marginTop = '0';
-      const sub = helper('Fill in the below to bring your Avatar to life. You can edit them later in Finances/Settings menus.');
-
-      const incomeAmt = input('Income Amount (£)', 'number', 'qsIncomeAmt', { min:'0', step:'0.01', placeholder:'e.g. 3200.00' });
-      const incomeCad = select('Income Cadence', 'qsIncomeCad', [['monthly','Monthly'],['weekly','Weekly'],['daily','Daily']]);
-      const expAmt    = input('Core Expenses Amount (£)', 'number', 'qsExpAmt', { min:'0', step:'0.01', placeholder:'e.g. 1800.00' });
-      const expCad    = select('Core Expenses Cadence', 'qsExpCad', [['monthly','Monthly'],['weekly','Weekly'],['daily','Daily']]);
-      const modeSel   = select('Start Mode', 'qsVitalsMode', [
-        ['accelerated','Standard'], ['safe','Cautious'], ['manual','Manual'],
-      ]);
-
-      const manualWrap = document.createElement('div'); manualWrap.style.display = 'none';
-      const manualInfo = helper('Optional: if you already spent this month before joining, enter a one‑number total and (optionally) split it between Stamina and Mana.');
-      const preAmt   = input('Pre‑start spend this month (£)', 'number', 'qsPreAmt', { min:'0', step:'0.01', placeholder:'e.g. 350.00' });
-      const preSplit = (()=>{ const box=document.createElement('div'); box.className='field';
-        box.innerHTML = `
-          <label>Split (optional)</label>
-          <div class="row"><input id="qsPreStaminaPct" type="number" class="input" min="0" max="100" step="1" value="60" /><span class="helper">Stamina %</span></div>
-          <div class="row" style="margin-top:.5rem;"><input id="qsPreManaPct" type="number" class="input" min="0" max="100" step="1" value="40" /><span class="helper">Mana %</span></div>`;
-        return box; })();
-      manualWrap.append(manualInfo, preAmt, preSplit);
-
-      const errBox = document.createElement('div');
-      errBox.id = 'qsError'; errBox.className = 'helper';
-      Object.assign(errBox.style, { display: 'none', color: '#ff6b6b', marginTop: '4px' });
-
-      const actions = document.createElement('div');
-      Object.assign(actions.style, { display:'flex', gap:'.6rem', flexWrap:'wrap', marginTop:'8px' });
-      const btnSave = document.createElement('button'); btnSave.className='ws-btn ws-primary'; btnSave.textContent='Save & Continue';
-      const btnSkip = document.createElement('button'); btnSkip.className='ws-btn ws-ghost';   btnSkip.textContent='Skip for now';
-      [btnSave, btnSkip].forEach(btn=>{
-        Object.assign(btn.style, { padding: '.6rem .9rem', borderRadius: '12px',
-          border: '1px solid var(--line,#333)', background: 'rgba(255,255,255,.06)', cursor: 'pointer', color: '#fff' });
-      });
-      btnSave.style.background = 'linear-gradient(180deg, rgba(90,180,255,.25), rgba(90,180,255,.12))';
-
-      card.append(title, sub, incomeAmt, incomeCad, expAmt, expCad, modeSel, manualWrap, errBox, actions);
-      actions.append(btnSave, btnSkip); shell.append(card);
-
-      const applySmall = () => {
-        if (window.innerWidth <= 360) {
-          shell.style.padding='10px'; card.style.width='96vw'; card.style.padding='12px'; card.style.borderRadius='14px';
-          actions.style.flexDirection='column'; btnSave.style.width=btnSkip.style.width='100%';
-        } else {
-          actions.style.flexDirection='row'; btnSave.style.width=btnSkip.style.width='';
-        }
-      };
-      applySmall(); window.addEventListener('resize', applySmall, { passive: true });
-
-      modeSel.querySelector('select').addEventListener('change', (e)=>{
-        manualWrap.style.display = (String(e.target.value||'safe') === 'manual') ? '' : 'none';
-      });
-      incomeCad.querySelector('select').value = 'monthly';
-      expCad.querySelector('select').value = 'monthly';
-      modeSel.querySelector('select').value = 'safe';
-
-      btnSave.addEventListener('click', async ()=>{
-        btnSave.disabled = true;
-        const showErr = (msg)=>{ errBox.textContent=msg; errBox.style.display='block'; };
-        const clearErr= ()=>{ errBox.textContent=''; errBox.style.display='none'; };
-        try{
-          const userRef = doc(db, 'players', uid);
-          const incomeAmount  = Math.max(0, Number(incomeAmt.querySelector('input').value || 0));
-          const incomeCadence = String(incomeCad.querySelector('select').value || 'monthly');
-          const expAmount     = Math.max(0, Number(expAmt.querySelector('input').value || 0));
-          const expCadence    = String(expCad.querySelector('select').value || 'monthly');
-          const vitalsMode    = String(modeSel.querySelector('select').value || 'accelerated');
-
-          clearErr();
-          if (incomeAmount <= 0) { showErr('Income must be greater than 0.'); btnSave.disabled=false; incomeAmt.querySelector('input').focus(); return; }
-          if (expAmount >= incomeAmount) { showErr('Core expenses must be less than income.'); btnSave.disabled=false; expAmt.querySelector('input').focus(); return; }
-
-          await updateIncome(incomeAmount, incomeCadence);
-          await updateCoreExpenses(expAmount, expCadence);
-          await updateDoc(userRef, { vitalsMode });
-
-          if (vitalsMode === 'manual') {
-            const total = Math.max(0, Number(preAmt.querySelector('input')?.value || 0));
-            let spct = Math.max(0, Math.min(100, Number(document.getElementById('qsPreStaminaPct')?.value || 0)));
-            let mpct = Math.max(0, Math.min(100, Number(document.getElementById('qsPreManaPct')?.value || 0)));
-            let sum = spct + mpct; if (sum <= 0.0001) { spct = 60; mpct = 40; sum = 100; }
-            spct = spct/sum; mpct = mpct/sum;
-
-            let startMs = Date.now();
-            const pSnap = await getDoc(userRef);
-            if (pSnap.exists()) {
-              const raw = pSnap.data()?.startDate;
-              if (raw?.toMillis) startMs = raw.toMillis();
-              else if (raw instanceof Date) startMs = raw.getTime();
-              else if (typeof raw === 'number') startMs = raw;
-            }
-            const d0 = new Date(startMs);
-            const startMonthStartMs = new Date(d0.getFullYear(), d0.getMonth(), 1).getTime();
-
-            await setDoc(doc(db, `players/${uid}/manualSeed/meta`), {
-              startMonthStartMs, startDateMs: startMs, updatedAtMs: Date.now()
-            }, { merge: true });
-
-            if (total > 0) {
-              await setDoc(doc(db, `players/${uid}/manualSeed/openingSummary`), {
-                totalPrestartDiscretionary: total,
-                split: { staminaPct: spct, manaPct: mpct },
-                providedAtMs: Date.now()
-              }, { merge: true });
-            }
-          }
-
-          await updateDoc(userRef, { onboardedAt: serverTimestamp() });
-          shell.remove(); window.removeEventListener('resize', applySmall);
-          await initHUD(uid);
-        } catch(e){
-          showErr('Something went wrong saving your setup. Please try again.');
-          console.warn('[Quick Setup] save failed:', e);
-          btnSave.disabled=false;
-        }
-      });
-
-      btnSkip.addEventListener('click', async ()=>{
-        try { await updateDoc(doc(db, 'players', uid), { onboardedAt: serverTimestamp() }); } catch(e){ console.warn('[Quick Setup] skip failed:', e); }
-        shell.remove(); window.removeEventListener('resize', applySmall);
-      });
-
-      document.body.appendChild(shell);
-    }
-
-    async function showWelcomeThenMaybeSetup(uid) {
-      const userRef = doc(db, 'players', uid);
-      const snap = await getDoc(userRef);
-      const already = snap.exists() && !!snap.data()?.onboardedAt;
-      if (already) return;
-
-      const INTRO_HTML = `
-        <h1>Welcome to Project MyFi</h1>
-        <p class="lead">Thanks for joining us in our goal to turn managing your finances into an enjoyable experience — we look forward to your feedback as the game continues to grow! 😊</p>
-        <p class="lead">
-          <strong>Project MyFi</strong> is a ever-evolving world, where your real-life spending habits power your in-game journey.<br><br>
-          Reimagine budgeting and spending, seeing the impact of your day-to-day choices through the eyes of your <em>Avatar</em>:<br><br>
-          • <strong>Progress</strong> - by rising to the challenge of <em>Quests</em> that promote saving.<br>
-          • <strong>Customise</strong> - based on your own habits, goals and style.<br>
-          • <strong>Empower</strong> -  and push deeper into <em>The Badlands</em>, earning rewards with real-world impact.<br><br> 
-          Spend wisely, plan intentionally, and grow stronger over time, all whilst having fun with your friends.
-        </p>
-      `;
-      const QUICKSTART_HTML = `
-        <h1>Quick Start</h1>
-        <p class="lead">Five key steps to get started.</p>
-        <ol class="steps">
-          <li>📜 Set your income and core (must-pay) expenses in the <strong>Finances</strong> menu.</li>
-          <li>📜 Log transactions in the <strong>Finances</strong> menu—long-press to edit, before they lock after 1 hour.</li>
-          <li>🌀 Manage <strong>Vitals</strong>: Stamina, Mana, Health, and Essence from a Daily, Weekly or Monthly view</li>
-          <li>⚔️ Keep your avatar healthy and empowered for <strong>The Badlands</strong> (coming soon).</li>
-          <li>❔ Check out the Help menu to learn more.</li>
-        </ol>
-      `;
-
-      const shell = document.createElement('div');
-      shell.innerHTML = `
-        <div class="wcard">
-          <div class="wcontent">${INTRO_HTML}</div>
-          <div class="actions">
-            <button id="wo-quickstart" class="ws-btn ws-primary">Quick Start</button>
-            <button id="wo-open" class="ws-btn ws-accent">Open Vitals</button>
-          </div>
-        </div>
-      `;
-      Object.assign(shell.style, {
-        position: 'fixed', inset: '0', display: 'grid', placeItems: 'center',
-        padding: '16px', background: 'rgba(10,10,14,.65)', backdropFilter: 'blur(6px)',
-        zIndex: '9999', width: '100%', maxWidth: '100%', overflowX: 'hidden', boxSizing: 'border-box'
-      });
-      const card = shell.querySelector('.wcard');
-      Object.assign(card.style, {
-        width: 'min(520px, 92vw)', maxHeight: '88vh', overflow: 'auto',
-        background: 'rgba(22,22,28,.92)', border: '1px solid var(--line, #333)',
-        borderRadius: '16px', padding: '16px 16px 12px',
-        boxShadow: '0 18px 50px rgba(0,0,0,.35)', boxSizing: 'border-box'
-      });
-      const content = shell.querySelector('.wcontent');
-      const actions = card.querySelector('.actions');
-      Object.assign(actions.style, { display: 'flex', gap: '.6rem', flexWrap: 'wrap' });
-
-      const btns = card.querySelectorAll('.ws-btn');
-      btns.forEach(btn => {
-        Object.assign(btn.style, { padding: '.6rem .9rem', borderRadius: '12px',
-          border: '1px solid var(--line,#333)', background: 'rgba(255,255,255,.04)',
-          cursor: 'pointer', color: '#fff' });
-        if (btn.classList.contains('ws-primary')) {
-          btn.style.background = 'linear-gradient(180deg, rgba(90,180,255,.25), rgba(90,180,255,.12))';
-        }
-      });
-
-      const applySmallScreen = () => {
-        if (window.innerWidth <= 360) {
-          shell.style.padding='10px'; card.style.width='96vw'; card.style.padding='12px'; card.style.borderRadius='14px';
-          actions.style.flexDirection='column'; btns.forEach(b=>b.style.width='100%');
-        } else {
-          actions.style.flexDirection='row'; btns.forEach(b=>b.style.width='');
-        }
-        const s = content.querySelector('.steps'); if (s) Object.assign(s.style, {
-          margin: '0 0 12px 1.1rem', padding: '0', lineHeight: '1.35', fontSize: 'clamp(13px, 3.4vw, 15px)', overflowWrap: 'anywhere'
-        });
-      };
-      applySmallScreen(); window.addEventListener('resize', applySmallScreen, { passive: true });
-
-      const closeAndStartSetup = async () => {
-        shell.remove(); window.removeEventListener('resize', applySmallScreen);
-        await showFirstRunSetup(uid);
-      };
-
-      let mode = 'intro';
-      function render(next){ mode = next; content.innerHTML = (mode==='quick') ? QUICKSTART_HTML : INTRO_HTML;
-        const qb = shell.querySelector('#wo-quickstart'); if (qb) qb.textContent = (mode==='quick') ? 'Welcome' : 'Quick Start';
-        applySmallScreen();
-      }
-      shell.addEventListener('click', (e) => {
-        const btn = e.target.closest('#wo-open,#wo-quickstart'); if (!btn) return;
-        if (btn.id==='wo-open') { closeAndStartSetup(); return; }
-        if (btn.id==='wo-quickstart') { render(mode==='quick' ? 'intro' : 'quick'); return; }
-      });
-
-      render('intro'); document.body.appendChild(shell);
-    }
-
+    
     /* ---------- Load player + HUD ---------- */
     const uid = user.uid;
     window.localStorage.setItem('user', JSON.stringify(user));
+
+    // After: const uid = user.uid;
+    const userRef = doc(db, 'players', uid);
+
+    // Decide early if Welcome needs to show
+    let needsWelcome = false;
+    try {
+      const snap = await getDoc(userRef);
+      const welcomeDoneLocal  = localStorage.getItem('myfi.welcome.v1.done') === '1';
+      const welcomeDoneServer = snap.exists() && !!snap.data()?.onboarding?.welcomeDone;
+      needsWelcome = !(welcomeDoneLocal || welcomeDoneServer);
+    } catch (_) { /* if uncertain, default to not suppressing */ }
+
+    // If Welcome will show, suppress tours *before* HUD (and thus before Vitals tour tries to start)
+    if (needsWelcome) {
+      window.__MYFI_TOUR_SUPPRESS = true;
+    }
+    
 
     const vitalsPromise = (async () => {
       const userRef = doc(db, 'players', uid);
@@ -334,9 +109,9 @@ document.addEventListener("DOMContentLoaded", () => {
       await vitalsPromise;
     }
 
-
-
+    // ✅ Show welcome exactly once, after splash is done and HUD is ready
     await showWelcomeThenMaybeSetup(uid);
+
 
     /* ---------- SLIDE ROUTER (Vitals hub + satellites) ---------- */
     (function setupSlideRouter(){
